# public/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Masik Local Server REST API
  description: Masik Local Server RESTFULL API
  version: 1.0.0
servers:
  - url: http://10.12.1.11:4322/
    description: for dev online.
  - url: http://localhost:4323/
    description: for development.
paths:
  /api/v1/operate:
    post:
      deprecated: true
      operationId: operate
      tags:
        - OperateService
      summary: operate.
      description: operate.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                hid:
                  type: string
                id:
                  type: string
                method:
                  enum:
                    - Add
                    - Remove
                  type: string
                type:
                  enum: ["dcard", "duser", "mcard", "muser"]
                  type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /api/v1/get_operates:
    get:
      operationId: getOperates
      tags:
        - SyncService
      summary: get operates.
      description: get operates.
      security:
        - basicAuth: []
      parameters:
        - description: Hardware ID.
          name: hid
          in: query
          required: true
          schema:
            type: string
        - description: Start operate key.
          name: start_operate_key
          in: query
          required: true
          schema:
            format: int64
            type: string
        - description: End operate key.
          name: end_operate_key
          in: query
          schema:
            format: int64
            type: string
        - description: Type.
          name: type
          in: query
          schema:
            enum: ["dcard", "duser", "mcard", "muser"]
            type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /api/v1/latest_snapshot_operatekey:
    get:
      operationId: latestSnapshotOperatekey
      tags:
        - SyncService
      summary: get latest snapshot operatekey.
      description: get latest snapshot operatekey.
      security:
        - basicAuth: []
      parameters:
        - description: Hardware ID.
          name: hid
          in: query
          required: true
          schema:
            type: string
        - description: Type.
          name: type
          in: query
          schema:
            enum: ["dcard", "duser", "mcard", "muser"]
            type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /api/v1/snapshot_cards:
    get:
      operationId: snapshotCards
      tags:
        - SyncService
      summary: get snapshot cards.
      description: get snapshot cards.
      security:
        - basicAuth: []
      parameters:
        - description: Hardware ID.
          name: hid
          in: query
          required: true
          schema:
            type: string
        - description: Operate key.
          name: operate_key
          in: query
          required: true
          schema:
            format: int64
            type: string
        - description: Type.
          name: type
          in: query
          schema:
            enum: ["dcard", "duser", "mcard", "muser"]
            type: string
        - description: Page.
          name: page
          in: query
          schema:
            format: int64
            type: string
        - description: Size.
          name: size
          in: query
          schema:
            format: int64
            type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /api/v1/replace_card:
    put:
      operationId: replaceCard
      tags:
        - CardService
      summary: Replace an user's octopus card.
      description: Replace an user's octopus card.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                replacedCardID:
                  description: old card
                  type: string
                cardID:
                  description: new card
                  type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /api/v1/report:
    post:
      operationId: report
      tags:
        - CardService
      summary: Report data.
      description: Report data.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                time:
                  description: report time
                  type: number
                type:
                  description: |
                    | type  | 含义  |
                    | ------------ | ------------ |
                    | 1  | 门禁八达通  |
                    | 2  | 门禁User二维码, 本地验证  |
                    | 3  | 门禁qrcodeid二维码, 联网验证  |
                    | 4  | 门禁蓝牙  |
                    | 5  | 门禁远程开门  |
                    | 10  | 信箱上报状态  |
                  enum: [1, 2, 3, 4, 5, 10]
                  type: number
                customID:
                  description: device ID
                  type: string
                id:
                  description: |
                    for 1~5
                    - array: [x, x]
                    - string: x
                  oneOf:
                    - type: array
                      items:
                        type: string
                    - type: string
                result:
                  description: for 1~5
                  type: boolean
                lock:
                  description: for 10
                  type: array
                  items:
                    type: number
                ir:
                  description: for 10
                  type: array
                  items:
                    type: number
                groupID:
                  description: for 10
                  type: number
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /api/v1/qrcode_can_open_door:
    post:
      operationId: qrcodeCanOpenDoor
      tags:
        - QrcodeService
      summary: Qrcode can open door.
      description: Qrcode can open door.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qrcodeID:
                  type: string
                megaDeviceCustomID:
                  type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                properties: {}
  /ota/${category}/{filename}:
    get:
      operationId: getOta
      tags:
        - OtaService
      summary: Ota.
      description: Ota.
      parameters:
        - name: category
          in: path
          description: category.
          required: true
          schema:
            type: string
        - name: filename
          in: path
          description: file name.
          required: true
          schema:
            type: string
      responses:
        200:
          description: A successful response.
  /api/v1/lift/get_info:
    get:
      operationId: liftGetInfo
      tags:
        - LiftService
      summary: Get lift info.
      description: 获取电梯信息.
      security:
        - basicAuth: []
      parameters:
        - description: 呼梯客户编码.
          name: customID
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Lift info response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      canContrl:
                        description: 是否可控.
                        type: boolean
                      direction:
                        description: 呼梯是否按下的状态
                        type: array
                        items:
                          type: string
                          enum: ["Up", "Dowm"]
                      infos:
                        description: 电梯状态.
                        type: array
                        items:
                          type: object
                          properties:
                            customID:
                              description: 按梯梯箱客户编码(MA001)
                              type: string
                            canContrl:
                              description: 是否可控.
                              type: boolean
                            floor:
                              description: 楼层标识(G)
                              enum: ['-1', '0', 'G', '2', '3']
                              type: string
                            direction:
                              description: 方向
                              enum: ["Up", "Dowm"]
                              type: string
  /api/v1/lift/contrl_outside:
    put:
      operationId: liftContrlOutside
      tags:
        - LiftService
      summary: Contrl the lift from outside.
      description: 呼梯呼叫某一个电梯组的接口.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customID:
                  description: 呼梯客户编码(CAL001)
                  type: string
                floor:
                  description: 楼层标识(G)
                  enum: ['-1', '0', 'G', '2', '3']
                  type: string
                direction:
                  description: 方向
                  enum: ["Up", "Dowm"]
                  type: string
      responses:
        200:
          description: Contrl outside response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
  /api/v1/lift/contrl_inside:
    put:
      operationId: liftContrlInside
      tags:
        - LiftService
      summary: Contrl the lift from inside.
      description: Contrl the lift from inside.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customID:
                  description: 按梯梯箱客户编码(MA001)
                  type: string
                floor:
                  description: 楼层标识(G)
                  enum: ['-1', '0', 'G', '2', '3']
                  type: string
      responses:
        200:
          description: Contrl outside response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
components:
  securitySchemes:
    basicAuth: # <-- arbitrary name for the security scheme
      type: http
      scheme: basic